fix: Display backup frequency as readable label instead of cron expression
The Automatic Backup card on the Object Service dashboard was showing
the raw cron expression (e.g. "0 0 * * *"). This converts it to a
human-readable label (Daily, Weekly, Monthly) matching what users
selected during Storage System creation.
